code/requirements.txt
absl-py==2.1.0
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
bcc==0.18.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
dm-tree==0.1.8
docstring_parser==0.16
etils==1.10.0
exceptiongroup==1.2.2
executing==2.1.0
file-read-backwards==3.1.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
immutabledict==4.2.0
importlib-metadata==4.6.4
importlib_resources==6.4.5
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
jeepney==0.7.1
jupyterlab_widgets==3.0.13
kagglehub==0.3.4
keras==3.4.1
keras-core==0.1.7
keras-cv==0.9.0
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==8.10.0
mpmath==1.3.0
namex==0.0.8
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pop-transition==1.1.2
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2024.11.6
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.14.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
simple-parsing==0.1.6
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.13.2
systemd-python==234
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
toml==0.10.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
